+++ CHANGELOG.md
@@ -7,8 +7,6 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 
 ## [Unreleased]
 ### Added
-- Checks for structs that are not marked as read-only.
-- Checks for classes that are not marked as static, sealed or abstract.
 ### Fixed
 ### Changed
 ### Removed
@@ -17,6 +15,11 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 <!-- 
 Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
 -->
+## [1.3.0] - 2019-04-19
+### Added
+- Checks for structs that are not marked as read-only.
+- Checks for classes that are not marked as static, sealed or abstract.
+
 ## [1.2.1] - 2019-03-30
 - Fixed prohibition of ISqlServerDatabase.QueryArbitrarySqlAsync<>